Task #4101, #4102: Updated .env file, docker-compose.yml, added custom Exceptions and refactoring of the exception package, fixed minor bugs in UserService
