+++ .env
@@ -1,3 +1,4 @@
 POSTGRES_DATABASE=wise_back
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
+POSTGRES_URL=jdbc:postgresql://localhost:5432/wise_back
