+++ docker-compose.yml
@@ -1,11 +1,12 @@
 version: "3.1"
 services:
   db:
-    image: postgres
+    image: postgres:15
     restart: always
     ports:
       - "5432:5432"
-    env_file: .env
+    env_file:
+      - .env
     environment:
       POSTGRES_DATABASE: ${POSTGRES_DATABASE}
       POSTGRES_USERNAME: ${POSTGRES_USERNAME}
