+++ pom.xml
@@ -12,6 +12,7 @@
     <artifactId>wise</artifactId>
     <version>1.0.0</version>
     <name>wise-api</name>
+    <packaging>jar</packaging>
     <description>Backend api for project Wise</description>
     <properties>
         <java.version>17</java.version>
@@ -53,7 +54,7 @@
         <dependency>
             <groupId>com.google.oauth-client</groupId>
             <artifactId>google-oauth-client</artifactId>
-            <version>1.33.3</version>
+            <version>1.34.1</version>
         </dependency>
         <dependency>
             <groupId>com.google.api-client</groupId>
