+++ src/main/java/com/teamback/wise/exceptions/auth/ExpiredRefreshTokenException.java
@@ -1,0 +1,9 @@
+package com.teamback.wise.exceptions.auth;
+
+public class ExpiredRefreshTokenException extends RuntimeException {
+
+    public ExpiredRefreshTokenException(String message) {
+        super(message);
+    }
+
+}
