+++ src/main/java/com/teamback/wise/exceptions/auth/JWTInvalidException.java
@@ -1,4 +1,4 @@
-package com.teamback.wise.exceptions;
+package com.teamback.wise.exceptions.auth;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
