+++ src/main/java/com/teamback/wise/exceptions/user/UserNotAuthenticatedException.java
@@ -1,0 +1,9 @@
+package com.teamback.wise.exceptions.user;
+
+public class UserNotAuthenticatedException extends RuntimeException {
+
+    public UserNotAuthenticatedException(String message) {
+        super(message);
+    }
+
+}
