+++ src/main/java/com/teamback/wise/exceptions/user/UserNotFoundException.java
@@ -1,0 +1,9 @@
+package com.teamback.wise.exceptions.user;
+
+public class UserNotFoundException extends RuntimeException {
+
+    public UserNotFoundException(String message) {
+        super(message);
+    }
+
+}
