+++ src/main/java/com/teamback/wise/security/filters/ExceptionHandlerFilter.java
@@ -2,7 +2,7 @@ package com.teamback.wise.security.filters;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.teamback.wise.exceptions.JWTInvalidException;
+import com.teamback.wise.exceptions.auth.JWTInvalidException;
 import com.teamback.wise.exceptions.exceptionHandler.FilterErrorDetails;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
