+++ src/main/resources/application.properties
@@ -1,8 +1,8 @@
 server.servlet.context-path=/api/v1
 
-spring.datasource.url=jdbc:postgresql://localhost:5432/wise_back
-spring.datasource.username=postgres
-spring.datasource.password=postgres
+spring.datasource.url=${POSTGRES_URL}
+spring.datasource.username=${POSTGRES_USER}
+spring.datasource.password=${POSTGRES_PASSWORD}
 
 spring.jpa.properties.hibernate.default_schema=public
 
@@ -17,5 +17,4 @@ jwt.issuer=wise
 jwt.expiration-time-access-token-mn=10
 jwt.expiration-time-refresh-token-mn=30
 
-google.client-id=139884528446-3t89gmbf4hub7rgb15ll8v7ddekq302d.apps.googleusercontent.com
-google.client-secret=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
+google.client-id=228236225556-394g22j3dfrr0bp8gfjgrm99hvsjb5q0.apps.googleusercontent.com
